Better emulate CP/M for zexall
The zero page has a specific layout, including "LD HL, (0006h)"
being used to find the jump address of the high CP/M code.
Zexall uses this to set the SP to grow down, so need to
emulate this to avoid the stack stomping on zero page
as it was before.